+++ src/main/java/com/vnano/defenders/enchantment/ModEnchantments.java
@@ -10,6 +10,7 @@ public final class ModEnchantments {
     public static EnchantmentDefender FINESSE;
     public static EnchantmentDefender REFLEXES;
     public static EnchantmentDefender DEFLECTION;
+    public static EnchantmentDefender SIXTH_SENSE;
 
     public static int getLevel(EnchantmentDefender enchantment, ItemStack stack) {
         return enchantment != null && enchantment.isEnabled()